fix the package name after the refactoring
