+++ app/src/main/java/net/squanchy/schedule/view/EventItemView.java
@@ -11,8 +11,8 @@ import java.util.Locale;
 import net.squanchy.R;
 import net.squanchy.eventdetails.widget.ExperienceLevelIconView;
 import net.squanchy.schedule.domain.view.Event;
+import net.squanchy.search.model.Speaker;
 import net.squanchy.service.firebase.model.FirebaseSpeaker;
-import net.squanchy.search.speaker.view.Speaker;
 import net.squanchy.support.widget.CardLayout;
 
 public class EventItemView extends CardLayout {
