+++ app/src/main/java/net/squanchy/search/view/SpeakerViewHolder.java
@@ -1,4 +1,4 @@
-package net.squanchy.speaker.view;
+package net.squanchy.search.view;
 
 import android.support.annotation.Nullable;
 import android.support.v7.widget.RecyclerView;
@@ -7,7 +7,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import net.squanchy.R;
-import net.squanchy.search.speaker.view.Speaker;
+import net.squanchy.search.model.Speaker;
 
 public class SpeakerViewHolder extends RecyclerView.ViewHolder {
