+++ app/src/main/java/net/squanchy/speaker/SpeakerService.java
@@ -5,7 +5,7 @@ import java.util.Comparator;
 import java.util.List;
 
 import net.squanchy.service.firebase.FirebaseDbService;
-import net.squanchy.search.speaker.view.Speaker;
+import net.squanchy.search.model.Speaker;
 
 import io.reactivex.Observable;
