+++ app/src/main/res/layout/activity_speaker_list.xml
@@ -20,7 +20,7 @@
 
   </android.support.design.widget.AppBarLayout>
 
-  <net.squanchy.speaker.view.SpeakersView
+  <net.squanchy.search.view.SpeakersView
     android:id="@+id/speakersView"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
